// barq-system-frontend/src/sdk/types.gen.ts
export type type = 'soc' | 'cybersecurity' | 'grc';
// ...
export type EventAgendaItemEntity = {
  id: number;
  event_id: number;
  From: string;
  To: string;
  subject: string;
  description: string;
  event: EventEntity;
  created_at: string;
  updated_at: string;
};

export type EventEntity = {
  id: number;
  crmId: string;
  name: string;
  registration_StartDate: string;
  registration_EndDate: string;
  event_StartDate: string;
  event_EndDate: string;
  country: string;
  city: string;
  location: string;
  shortDescription: string;
  Latitude: string;
  Longitude: string;
  Solution: string;
  Service: string;
  description: string;
  CampaignType: string;
  agendaItems: Array<EventAgendaItemEntity>;
  event_joinus_form: Array<EventJoinusFormEntity>;
  created_at: string;
  updated_at: string;
};

export type EventJoinusFormEntity = {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  phone_number: string;
  phone_number_key: string;
  position: string;
  company_name: string;
  event_id: number;
  event: EventEntity;
  created_at: string;
  updated_at: string;
};
// ...
export type EventsPartnerEntity = {
  id: number;
  title: string;
  sub_title: string;
  logos: Array<MediaEntity>;
  events_partner_id_events_partner_translations: Array<EventsPartnerTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type EventsPartnerTranslationsEntity = {
  id: number;
  events_partner_id: number;
  events_partner_translation_events_partner?: EventsPartnerEntity | null;
  title: string;
  sub_title: string;
  language: 'ar' | 'en';
};

export type EventsSpeakersCardsEntity = {
  id: number;
  name: string;
  role: string;
  image_id: number;
  events_speakers_id: number;
  image: MediaEntity;
  events_speakers_cards_id_events_speakers_cards_translations: Array<EventsSpeakersCardsTranslationsEntity>;
  events_speakers_cards_parent: EventsSpeakersEntity;
  created_at: string;
  updated_at: string;
};

export type EventsSpeakersCardsTranslationsEntity = {
  id: number;
  events_speakers_cards_id: number;
  events_speakers_cards_translation_events_speakers_cards?: EventsSpeakersCardsEntity | null;
  name: string;
  role: string;
  language: 'ar' | 'en';
};

export type EventsSpeakersEntity = {
  id: number;
  title: string;
  events_speakers_id_events_speakers_translations: Array<EventsSpeakersTranslationsEntity>;
  events_speakers_cards_id_events_speakers_cards: Array<EventsSpeakersCardsEntity>;
  created_at: string;
  updated_at: string;
};

export type EventsSpeakersTranslationsEntity = {
  id: number;
  events_speakers_id: number;
  events_speakers_translation_events_speakers?: EventsSpeakersEntity | null;
  title: string;
  language: 'ar' | 'en';
};
// ...
export type id = 'ASC' | 'DESC';

export type title = 'ASC' | 'DESC';

export type description = 'ASC' | 'DESC';

export type image_id = 'ASC' | 'DESC';

export type created_at = 'ASC' | 'DESC';

export type updated_at = 'ASC' | 'DESC';
// ...
export type sub_title = 'ASC' | 'DESC';
// ...
export type $op =
  | 'Eq'
  | 'NotEq'
  | 'GreaterThan'
  | 'LessThan'
  | 'GreaterThanOrEq'
  | 'LessThanOrEq'
  | 'IsNull'
  | 'IsNotNull'
  | 'In'
  | 'NotIn';
// ...
export type name = 'asc' | 'desc';
// ...
export type crmId = 'asc' | 'desc';

export type CampaignType = 'asc' | 'desc';

export type shortDescription = 'asc' | 'desc';
// ...
export type ReadEventFilters = {
  id?: {
    $val?: (string | number) | Array<string | number>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  crmId?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  name?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  registration_StartDate?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  registration_EndDate?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  event_StartDate?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  event_EndDate?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  country?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  city?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  location?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  shortDescription?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  Latitude?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  Longitude?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  Solution?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  Service?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  description?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  CampaignType?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  created_at?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  updated_at?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
};
// ...
export type ReadEventOrders = {
  id?: 'asc' | 'desc';
  crmId?: 'asc' | 'desc';
  name?: 'asc' | 'desc';
  registration_StartDate?: 'asc' | 'desc';
  registration_EndDate?: 'asc' | 'desc';
  event_StartDate?: 'asc' | 'desc';
  event_EndDate?: 'asc' | 'desc';
  country?: 'asc' | 'desc';
  city?: 'asc' | 'desc';
  location?: 'asc' | 'desc';
  shortDescription?: 'asc' | 'desc';
  Latitude?: 'asc' | 'desc';
  Longitude?: 'asc' | 'desc';
  Solution?: 'asc' | 'desc';
  Service?: 'asc' | 'desc';
  description?: 'asc' | 'desc';
  created_at?: 'asc' | 'desc';
  updated_at?: 'asc' | 'desc';
};

export type registration_StartDate = 'asc' | 'desc';

export type registration_EndDate = 'asc' | 'desc';

export type event_StartDate = 'asc' | 'desc';

export type event_EndDate = 'asc' | 'desc';

export type country = 'asc' | 'desc';

export type city = 'asc' | 'desc';

export type location = 'asc' | 'desc';

export type Latitude = 'asc' | 'desc';

export type Longitude = 'asc' | 'desc';

export type Solution = 'asc' | 'desc';

export type Service = 'asc' | 'desc';

export type ReadEventQuery = {
  filters?: ReadEventFilters;
  orders?: ReadEventOrders;
  relations?: {
    agendaItems?: boolean | string;
    event_joinus_form?: boolean | string;
  };
  pagination?: {
    skip?: string | number;
    take?: string | number;
  };
};

export type ReadEventRelations = {
  agendaItems?: boolean | string;
  event_joinus_form?: boolean | string;
};

export type ReadEventsPartnerFilters = {
  id?: {
    $val?: (string | number) | Array<string | number>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  title?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  sub_title?: {
    $val?: string | Array<string>;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'Contains'
      | 'StartsWith'
      | 'EndsWith'
      | 'IsNull'
      | 'IsNotNull'
      | 'In'
      | 'NotIn';
  } | null;
  created_at?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
  updated_at?: {
    $val?: string;
    $op:
      | 'Eq'
      | 'NotEq'
      | 'GreaterThan'
      | 'LessThan'
      | 'GreaterThanOrEq'
      | 'LessThanOrEq'
      | 'IsNull'
      | 'IsNotNull';
  } | null;
};

export type ReadEventsPartnerOrders = {
  id?: 'asc' | 'desc';
  title?: 'asc' | 'desc';
  sub_title?: 'asc' | 'desc';
  created_at?: 'asc' | 'desc';
  updated_at?: 'asc' | 'desc';
};

export type ReadEventsPartnerQuery = {
  filters?: ReadEventsPartnerFilters;
  orders?: ReadEventsPartnerOrders;
  relations?: {
    events_partner_id_events_partner_translations?: boolean | string;
    logos?: (string | boolean) | ReadMediaRelations;
  };
  pagination?: {
    skip?: string | number;
    take?: string | number;
  };
};

export type ReadEventsPartnerRelations = {
  events_partner_id_events_partner_translations?: boolean | string;
  logos?: (string | boolean) | ReadMediaRelations;
};

export type ReadEventsSpeakersFilters = {
  id?: number;
  title?: string;
  created_at?: string;
  updated_at?: string;
};

export type ReadEventsSpeakersOrders = {
  id?: 'ASC' | 'DESC';
  title?: 'ASC' | 'DESC';
  created_at?: 'ASC' | 'DESC';
  updated_at?: 'ASC' | 'DESC';
};

export type ReadEventsSpeakersQuery = {
  filters?: ReadEventsSpeakersFilters;
  orders?: ReadEventsSpeakersOrders;
  relations?: {
    events_speakers_id_events_speakers_translations?: boolean | string;
    events_speakers_cards_id_events_speakers_cards?:
      | (string | boolean)
      | {
          events_speakers_cards_id_events_speakers_cards_translations?:
            | boolean
            | string;
          image?: boolean | string;
        };
  };
  pagination?: {
    skip?: string | number;
    take?: string | number;
  };
};
// ...
export type number = 'asc' | 'desc';
// ...
export type key = 'asc' | 'desc';

export type size = 'asc' | 'desc';

export type mime_type = 'asc' | 'desc';

export type format = 'asc' | 'desc';

export type ReadMediaQuery = {
  filters?: ReadMediaFilters;
  relations?: {
    [key: string]: unknown;
  };
  orders?: ReadMediaOrders;
  pagination?: {
    skip?: string | number;
    take?: string | number;
  };
};

export type ReadMediaRelations = {
  [key: string]: unknown;
};
// ...
export type EventControllerReadResponse = {
  data?: Array<EventEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type EventsSpeakersControllerReadResponse = {
  data?: Array<EventsSpeakersEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type EventsPartnerControllerReadResponse = {
  data?: Array<EventsPartnerEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};